refactor: separate out multiple files for wizard
